Restructure navbar, main elements
to ensure `<header role="banner">` is present around logo, `<nav>` is just around navigational elements, and `<main>` wraps only main content.
