+++ scripts/dev/Vagrantfile
@@ -31,7 +31,7 @@ Vagrant.configure("2") do |config|
 
   config.vm.provider "virtualbox" do |v|
     # see https://stefanwrobel.com/how-to-make-vagrant-performance-not-suck
-    v.customize ["modifyvm", :id, "--memory", 3072]
+    v.customize ["modifyvm", :id, "--memory", 8192]
     # see https://www.mkwd.net/improve-vagrant-performance/
     v.cpus = 2
     v.customize ["modifyvm", :id, "--ioapic", "on"]
